+++ src/app/loadout-builder/process-worker/process-utils.ts
